fix: address type errors in define(...) call type assertion
The define(...) call types say a `number` is returned for u64s, when they
should actually be `bigint`. Since this obviously won't work (a u64 is
bigger than the largest `Number.MAX_SAFE_INTEGER`), I'm pretty sure this
has got to be an error in the typings.

I think this is probably worth digging more into at some point
(ie, is this actually a thing?), because if it isn't an error in
the typings, then it's possible there are cases where there would be data
loss sometimes due to truncation.