Fix lots of minor type errors (#14)
* fix: address type errors in platform lookup

* fix: address type errors in define(...) call type assertion

The define(...) call types say a `number` is returned for u64s, when they
should actually be `bigint`. Since this obviously won't work (a u64 is
bigger than the largest `Number.MAX_SAFE_INTEGER`), I'm pretty sure this
has got to be an error in the typings.

I think this is probably worth digging more into at some point
(ie, is this actually a thing?), because if it isn't an error in
the typings, then it's possible there are cases where there would be data
loss sometimes due to truncation.

* fix: address s/listError/liftError/ typo

* feat: make liftError return a [enumname, variant] pair so that it can be fed into UniffiError

I think I must be missing something from the react native bindgen, I
think it handles this in a much more elegant way where the error lift
step actually returns an Error subclass.

Anyway, this is also something else to look into longer term.

* fix: add back in missing asyncopts

* fix: ignore FinalizationRegistry type errors

This isn't defined on all platforms and typescript complains, however
there's a guard for this in the code so the type error produced here
isn't all that actionable.